Initial code dump
This is a mostly working, manually tested version of pip-stale. It's
based on a script I wrote after piprot stopped working.

The big feature here is that this supports figuring out the latest
version, latest minor version, and latest patch and it works with
requirements files, packages passed in via the command line, and
environments.