Overhaul requirements parsing (#8, #9)
This overhauls requirements parsing removing our parser in favor of the
packaging library parser. That fixes the issue pip-stale had in
issue #8. This adds tests for requirements parsing.

This also changes the table so it uses box.MARKDOWN which is more
copy/paste friendly. That's issue #9.

This also adds a verbose mode.