Add --show, overhaul output, add more tests
This adds a --show argument that lets you specify which latest versions
to show (latest, minor, patch).

It fixes the table output format so it doesn't show blanks anymore--that
seemed too weird.

It adds a csv output format.

It adds a bunch of tests for showing output and parsing requirements
files.